Make e2e work ,add Razzle.cmd for quick start and fix some bugs

+++ DocContextMenuPackage/DocContextMenuPackage/MDGenerateHelper.cs
@@ -28,17 +28,24 @@ namespace Company.DocContextMenuPackage
             string docMetadataPath = Path.GetDirectoryName(_docsFolder.Properties.Item("FullPath").Value.ToString());
             string arguments = string.Format("\"{0}\" /o:\"{1}\" /p:\"{2}\" /t:\"Markdown\"", _dte.Solution.FullName, docMetadataPath, projectName);
 
-            await ProcessUtility.ExecuteWin32ProcessAsync(executorPath, arguments, workingDirectory, 100000);
+            var processingDetail = await ProcessUtility.ExecuteWin32ProcessAsync(executorPath, arguments, workingDirectory, 100000);
+
+            if (processingDetail.ExitCode != 0)
+            {
+                throw new System.ApplicationException(string.Format("Error executing {0} {1} : {2}", executorPath, arguments, processingDetail.StandardOutput + processingDetail.StandardError));
+            }
+
             foreach (string file in Directory.GetFiles(docMetadataPath + "\\" + _project.Name))
             {
                File.Delete(file);
             }
 
             int itemCount = _projectDocFolder.ProjectItems.Count;
-            for (int i=itemCount; i>=1; i--)
+            for (int i = itemCount; i >= 1; i--)
             {
-               _projectDocFolder.ProjectItems.Item(i).Remove();
+                _projectDocFolder.ProjectItems.Item(i).Remove();
             }
+
             foreach (string dir in Directory.GetDirectories(docMetadataPath + "\\mdtoc\\" + _project.Name))
             {
                 foreach (string file in Directory.GetFiles(dir))
